added test for WordProcessor and re-arranged the main test file into independent tests
